requirements.txt
flask~=3.0.0
werkzeug~=3.1.3
gunicorn==21.2.0
Jinja2==3.1.2
psycopg2-binary
